Formulating compression option, generalizing genemapAPMI; MULTIPLE NOTES BELOW
Compression of all strings (regulators or target names) is intended to greatly reduce memory load of the entire program, as large dictionaries will have to contain identifiers for each regulator->target, which can reach numbers on the order of 50 million edges.  As a string, this is an inefficient data structure.  Compression will be the default method of running ARACNe3, but right now it is in early stages.  Next steps include making several functions generalizable to the bucket types of genemap, as we can have genemap: string -> vector<float> and genemap: uint16_t -> vector<float>, as the compression option.

Once we have compression modularity, I will re-evaluate the APMI functions to instead deal with a new data structure that will greatly improve the efficiency of *both* the first and second pruning steps.  Namely, we will have a dictionary of regulator identifiers as the key, and the value will be a vector of a ball-on-stick type of struct, that contains the target, MI, and in the case of first pruning step, p-value.  This is the most condensed way of storing this information explored thus far.

Once the null model can be entirely finished, the first pruning step will be complete, in a very efficient form.

For the time being, I am dropping any multithreading options, and they will be incorporated using the C++ standard multithreading library, as soon as we can run the entire ARACNe3 on just one thread.